refactor: Openai video model 移动到 dto
