+++ dto/openai_video.go
@@ -1,4 +1,4 @@
-package common
+package dto
 
 import (
 	"strconv"
