+++ relay/channel/adapter.go
@@ -53,5 +53,5 @@ type TaskAdaptor interface {
 }
 
 type OpenAIVideoConverter interface {
-	ConvertToOpenAIVideo(originTask *model.Task) (*relaycommon.OpenAIVideo, error)
+	ConvertToOpenAIVideo(originTask *model.Task) (*dto.OpenAIVideo, error)
 }
