+++ relay/channel/task/sora/adaptor.go
@@ -186,8 +186,8 @@ func (a *TaskAdaptor) ParseTaskResult(respBody []byte) (*relaycommon.TaskInfo, e
 	return &taskResult, nil
 }
 
-func (a *TaskAdaptor) ConvertToOpenAIVideo(task *model.Task) (*relaycommon.OpenAIVideo, error) {
-	openAIVideo := &relaycommon.OpenAIVideo{}
+func (a *TaskAdaptor) ConvertToOpenAIVideo(task *model.Task) (*dto.OpenAIVideo, error) {
+	openAIVideo := &dto.OpenAIVideo{}
 	err := json.Unmarshal(task.Data, openAIVideo)
 	if err != nil {
 		return nil, errors.Wrap(err, "unmarshal to OpenAIVideo failed")
